# app/public/openapi.yaml
openapi: 3.1.0
info:
  title: ClawdVault API
  description: |
    API for ClawdVault - a pump.fun-style token launchpad on Solana.
    
    ## 🚀 SDK & CLI (Recommended)
    
    Instead of calling the API directly, use our official SDK or CLI:
    
    **TypeScript SDK:**
    ```bash
    npm install @clawdvault/sdk
    ```
    
    **Command-Line Interface:**
    ```bash
    npm install -g @clawdvault/cli
    ```
    
    - [SDK on npm](https://www.npmjs.com/package/@clawdvault/sdk)
    - [CLI on npm](https://www.npmjs.com/package/@clawdvault/cli)
    - [GitHub](https://github.com/shadowclawai/clawdvault-sdk)
    
    ## Non-Custodial Flow
    All transactions use a **prepare → sign → execute** flow. Your private key never leaves your device.
    
    ## Bonding Curve
    - Initial: 30 SOL / 1.073B tokens
    - Fee: 1% (0.5% protocol + 0.5% creator)
    - Graduation: 120 SOL raised → auto-migrates to Raydium
    - Post-graduation: ~0.25% Raydium fee, trades via Jupiter
  version: 1.0.0
  contact:
    name: ClawdVault
    url: https://clawdvault.com
  x-sdk:
    name: "@clawdvault/sdk"
    url: https://www.npmjs.com/package/@clawdvault/sdk
    github: https://github.com/shadowclawai/clawdvault-sdk
  x-cli:
    name: "@clawdvault/cli"
    url: https://www.npmjs.com/package/@clawdvault/cli
servers:
  - url: https://clawdvault.com/api
    description: Production (Mainnet)
...
paths:
...
  /token/execute-create:
    post:
      tags: [Tokens]
      summary: Execute token creation
      description: "Step 2: Submit the signed transaction and create database record."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [signedTransaction, mint, creator, name, symbol]
              properties:
                signedTransaction:
                  type: string
                  description: Base64 signed transaction
                mint:
                  type: string
                creator:
                  type: string
                name:
                  type: string
                symbol:
                  type: string
                description:
                  type: string
                image:
                  type: string
                  format: uri
                twitter:
                  type: string
                telegram:
                  type: string
                website:
                  type: string
                # NOTE: initialBuy removed - contract emits TradeEvent, sync-trades catches it
      responses:
        '200':
          description: Token created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  signature:
                    type: string
                  mint:
                    type: string
                  token:
                    $ref: '#/components/schemas/Token'
                  # NOTE: initialBuyTrade removed - sync-trades handles initial buy recording
                  explorer:
                    type: string
                    format: uri
...
  /trades:
    get:
      tags: [Price Data]
      summary: Get trade history
      parameters:
        - name: mint
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: before
          in: query
          schema:
            type: string
            format: date-time
          description: Cursor for pagination
      responses:
        '200':
          description: Trade history
          content:
            application/json:
              schema:
                type: object
                properties:
                  trades:
                    type: array
                    items:
                      $ref: '#/components/schemas/Trade'
...
  /profile:
    get:
      tags: [Users]
      summary: Get user profile
      parameters:
        - name: wallet
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
    post:
      tags: [Users]
      summary: Update profile
      security:
        - WalletSignature: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  minLength: 2
                  maxLength: 20
                  pattern: '^[a-zA-Z0-9_-]+$'
                avatar:
                  type: string
                  format: uri
      responses:
        '200':
          description: Profile updated
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: JWT session token from /auth/session
    WalletSignature:
      type: apiKey
      in: header
      name: X-Wallet
      description: |
        Wallet address in X-Wallet header + signature in X-Signature header.
        Sign the request body (or action object) with your wallet.
  schemas:
    Token:
      type: object
      properties:
        mint:
          type: string
        name:
          type: string
        symbol:
          type: string
        description:
          type: string
        image:
          type: string
        creator:
          type: string
        creator_name:
          type: string
        price_sol:
          type: number
        price_usd:
          type: number
          description: Price in USD
        market_cap_sol:
          type: number
        market_cap_usd:
          type: number
          description: Market cap in USD
        volume_24h:
          type: number
        trades_24h:
          type: integer
          description: Number of trades in last 24 hours
        holders:
          type: integer
          description: Number of unique holders
        price_change_24h:
          type: number
          nullable: true
          description: 24-hour price change percentage (computed from price_24h_ago)
        ath:
          type: number
          description: All-time high price in USD
        price_24h_ago:
          type: number
          description: USD price 24 hours ago (for realtime % change calculation)
        virtual_sol_reserves:
          type: number
        virtual_token_reserves:
          type: number
        real_sol_reserves:
          type: number
        real_token_reserves:
          type: number
        graduated:
          type: boolean
        raydium_pool:
          type: string
          description: Raydium pool address (set after graduation)
        twitter:
          type: string
        telegram:
          type: string
        website:
          type: string
        last_trade_at:
          type: string
          format: date-time
          description: Timestamp of most recent trade
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Trade:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [buy, sell]
        sol_amount:
          type: number
        token_amount:
          type: number
        price_sol:
          type: number
          description: Price in SOL
        price_usd:
          type: number
          nullable: true
          description: Price in USD (calculated from sol_price_usd)
        sol_price_usd:
          type: number
          nullable: true
          description: SOL price at time of trade
        trader:
          type: string
        username:
          type: string
          nullable: true
          description: Trader's username if profile exists
        signature:
          type: string
        created_at:
          type: string
          format: date-time
...
    UserProfile:
      type: object
      properties:
        wallet:
          type: string
        username:
          type: string
        avatar:
          type: string
        created_at:
          type: string
          format: date-time
...